feat: enhance product management with improved validation and logging
- Added debug logging for request and file upload data in the postStore method.
- Updated validation error handling to provide more detailed feedback.
- Changed the 'harga' field type from decimal to float in the Produk model for better precision.
- Enhanced the create product view to handle category selection and display appropriate messages.
- Improved JavaScript for price input formatting and validation before form submission.

+++ app/Models/Produk.php
@@ -24,7 +24,7 @@ class Produk extends Model
     protected bool $updateOnlyChanged = true;
 
     protected array $casts = [
-        'harga' => 'decimal',
+        'harga' => 'float',
         'stok' => 'integer',
         'kategori_id' => 'integer'
     ];
